backend/Md.Tga.IntegrationTest/requirements.txt
google-cloud-firestore==2.4.0
google-cloud-pubsub==2.12.0
